fix: Potential fix for 3785
